requirements.txt
BeautifulSoup==3.2.0
MySQL-python==1.2.3
pillow==3.1.0
amqplib==1.0.0
anyjson==0.3.1
boto==2.8.0
celery==2.2.7
distribute==0.7.3
kombu==1.2.1
pycurl==7.19.0
pyparsing==1.5.6
python-dateutil==1.5
python-postmark==0.3.1
recaptcha-client==1.0.6
tornado==4.0.2
torndb==0.3
tornadotoad==0.3
tweepy==2.3
wsgiref==0.1.2
requests==1.2.3
mock==1.0.1
stripe==1.28.0
